refactor: extract Provider.Assignments
Move program-staff assignment commands, the five active-assignment reads,
and their helpers onto Provider.Assignments. Cross-domain get_staff_member
routes through the facade; dispatch pins the bus context to KlassHero.Provider
(not the sub-module) to preserve the durable event topic.
